refactoring...#1 refactored commands. - not yet ready, to do next:     - refactor 'events' modules content
