+++ core/services/bad_words.py
