+++ core/services/gemini.py
