+++ web-api/proj/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for tact_api project.
+Django settings for proj project.
 
 Generated by 'django-admin startproject' using Django 5.1.5.
 
@@ -49,7 +49,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'tact_api.urls'
+ROOT_URLCONF = 'proj.urls'
 
 TEMPLATES = [
     {
@@ -67,7 +67,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'tact_api.wsgi.application'
+WSGI_APPLICATION = 'proj.wsgi.application'
 
 
 # Database
