+++ web-api/proj/urls.py
@@ -1,5 +1,5 @@
 """
-URL configuration for tact_api project.
+URL configuration for proj project.
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/5.1/topics/http/urls/
